refactors kubeclient
