+++ cmd/run-server.go
@@ -2,15 +2,13 @@ package main
 
 import (
 	"sighupio/permission-manager/internal/config"
-	"sighupio/permission-manager/internal/kubeclient"
 	"sighupio/permission-manager/internal/server"
 )
 
 func main() {
 	cfg := config.New()
-	kc := kubeclient.New()
 
-	s := server.New(kc, *cfg)
+	s := server.New(*cfg)
 	addr := ":" + cfg.Backend.Port
 	s.Logger.Fatal(s.Start(addr))
 }
